% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aov_HLclass.R
\name{aov_HLclass}
\alias{aov_HLclass}
\title{This function is used to run an ANOVA.}
\usage{
aov_HLclass(data)
}
\arguments{
\item{data}{the data set being used, should be .csv}
}
\value{
ANOVA results in a summary table including total and residual Degrees of freedom, Sum of Squares, Mean Squares, F-statistic, and P-value.
}
\description{
A specific ANOVA that compares the reclassified alligators size classes to their HL ratios.
}
