% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Kmeans_Cluster.R
\name{cluster_data}
\alias{cluster_data}
\title{This function uses K-means cluster to create cluster groups based on geographic location.}
\usage{
cluster_data(data, lat_column, long_column, k)
}
\arguments{
\item{data}{the data set being used, should be .csv, and should have separate columns with latitude and longitude data.}

\item{lat_column}{The column name of your latitude data.}

\item{long_column}{The column name of your longitude data.}

\item{K}{is the number of groups you wish to generate, i.e., K=10 generates ten cluster groups to house observations in.}
}
\value{
a mutated data set with a four new columns: cluster group, animals per cluster, latitude center, and longitude center.
}
\description{
This function adds four new columns to your original data set.
These new columns can be utilized to conducted additional spatial analysis.
}
