% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Cleveland_plot.R
\name{plot_LLRatio_by_class}
\alias{plot_LLRatio_by_class}
\title{This function creates a Cleveland plot.}
\usage{
plot_LLRatio_by_class(data, x_axis, y_axis, binwidth)
}
\arguments{
\item{data}{the data set being used, should be .csv.}

\item{x_axis}{this is the column data that will be used for the x axis, this should be your classification data.}

\item{y_axis}{this is the column data that will be used for the y axis, this should be your continuous leukocyte:length ratio data.}

\item{binwidth}{sets the binwidth/dot size. This should be a numeric value. Suggested range is from 0.0020 - 0.0030 but your preference may be outside this range.}
}
\value{
A Cleveland plot by classification of leukocyte:length ratio data.
}
\description{
This plot compares the leukocyte:length ratio by animal classification.
}
